[svn r19422] less field access

+++ src/edu/sc/seis/fissuresUtil/hibernate/NetworkDB.java
@@ -64,14 +64,14 @@ public class NetworkDB extends AbstractHibernateDB {
     }
 
     public int put(ChannelImpl chan) {
-        Integer dbid;
+        Long dbid;
         internUnit(chan);
         try {
             chan.my_site.my_station = getStationById(chan.my_site.my_station.get_id());
         } catch(NotFound e) {
             int staDbid = put((StationImpl)chan.my_site.my_station);
         }
-        dbid = (Integer)getSession().save(chan);
+        dbid = (Long)getSession().save(chan);
         return dbid.intValue();
     }
